requirements.txt
pytest>=2.9.2
pyasn1>=0.1.9
pysnmp>=4.3.2
jira>=1.0.3
Twisted==16.2.0
pexpect==4.1.0
enum34>=1.1.6
jsonrpclib-pelix>=0.2.7
netaddr==0.7.18
paramiko>=2.0.1
cryptography==1.2.1
py2chainmap>=0.1.0
pytest-xdist>=1.11
pycurl>=7.43.0
PyYAML==3.11
Pygments>=2.1.3
-e git+https://github.com/mike01/pypacker@v3.3#egg=pypacker
